refactor(web): modularize RelatedBusinessesSection into card and sidebar sub-components
