+++ apps/web/src/components/user/related-businesses/RelatedBusinessesSection.tsx
@@ -1,27 +1,19 @@
 "use client";
 
 import { useMemo, useRef } from "react";
-import Link from "next/link";
 import { useQuery } from "@tanstack/react-query";
-import { AlertCircle, ChevronLeft, ChevronRight, MapPin, RefreshCcw, Wrench } from "@/icons/IconRegistry";
+import { AlertCircle, ChevronLeft, ChevronRight, RefreshCcw } from "@/icons/IconRegistry";
 
-import { getBusinesses, type Business } from "@/lib/api/user/businesses";
+import { getBusinesses } from "@/lib/api/user/businesses";
 import type { UserPage } from "@/lib/routeUtils";
-import {
-  DEFAULT_IMAGE_PLACEHOLDER,
-  toSafeImageSrc,
-} from "@/lib/image/imageUrl";
 import {
   type RelatedBusinessesDiscoveryContext,
   normalizeRelatedBusinessesDiscoveryContext,
 } from "@/lib/listings/listingDiscoveryContext";
-import { resolveListingLocationLabel } from "@/lib/listings/listingPresentation";
-import { Card } from "@/components/ui/card";
-import { Badge } from "@/components/ui/badge";
 import { Button } from "@esparex/ui";
 import { queryKeys } from "@/hooks/queries/queryKeys";
-
-import { SafeImage } from "@/components/ui/SafeImage";
+import { RelatedBusinessCard } from "./RelatedBusinessCard";
+import { RelatedBusinessSidebar } from "./RelatedBusinessSidebar";
 
 interface RelatedBusinessesSectionProps {
   context: RelatedBusinessesDiscoveryContext;
@@ -99,197 +91,26 @@ export function RelatedBusinessesSection({
 
   const scrollCarousel = (direction: "left" | "right") => {
     if (!carouselRef.current) return;
-    const scrollAmount = variant === "sidebar" ? 180 : 320;
-    if (variant === "sidebar") {
-      carouselRef.current.scrollBy({
-        top: direction === "left" ? -scrollAmount : scrollAmount,
-        behavior: "smooth",
-      });
-    } else {
-      carouselRef.current.scrollBy({
-        left: direction === "left" ? -scrollAmount : scrollAmount,
-        behavior: "smooth",
-      });
-    }
+    const scrollAmount = 320;
+    carouselRef.current.scrollBy({
+      left: direction === "left" ? -scrollAmount : scrollAmount,
+      behavior: "smooth",
+    });
   };
 
   if (variant === "sidebar") {
     return (
-      <div className="space-y-3 pt-2">
-        <div className="flex items-start justify-between gap-2">
-          <div>
-            <h3 className="text-sm font-bold text-slate-800">{sectionCopy.title}</h3>
-            <p className="text-2xs text-slate-500 line-clamp-1">{sectionCopy.description}</p>
-          </div>
-          {!isLoading && businesses.length > 0 && (
-            <div className="flex items-center gap-1 shrink-0">
-              <Button
-                size="icon"
-                variant="outline"
-                className="h-7 w-7 rounded-lg border-slate-200"
-                onClick={() => scrollCarousel("left")}
-                aria-label="Previous services"
-              >
-                <ChevronLeft className="h-3.5 w-3.5" />
-              </Button>
-              <Button
-                size="icon"
-                variant="outline"
-                className="h-7 w-7 rounded-lg border-slate-200"
-                onClick={() => scrollCarousel("right")}
-                aria-label="Next services"
-              >
-                <ChevronRight className="h-3.5 w-3.5" />
-              </Button>
-            </div>
-          )}
-        </div>
-
-        {isLoading ? (
-          <div className="space-y-2">
-            {Array.from({ length: 3 }).map((_, index) => (
-              <div key={index} className="h-14 w-full animate-pulse rounded-xl bg-slate-100" />
-            ))}
-          </div>
-        ) : null}
-
-        {!isLoading && isError ? (
-          <div className="rounded-xl border border-amber-200 bg-amber-50 p-2.5 text-xs text-amber-800">
-            <span>Unable to load nearby services.</span>
-          </div>
-        ) : null}
-
-        {!isLoading && !isError && businesses.length === 0 ? (
-          <div className="rounded-xl border border-slate-200 bg-slate-50 p-2.5 text-2xs text-slate-500">
-            {sectionCopy.empty}
-          </div>
-        ) : null}
-
-        {!isLoading && !isError && businesses.length > 0 ? (
-          <div
-            ref={carouselRef}
-            className="flex flex-col gap-2 max-h-[220px] overflow-y-auto scrollbar-hide scroll-smooth"
-            style={{ scrollbarWidth: "none", msOverflowStyle: "none" }}
-          >
-            {businesses.map((business) => {
-              const distanceLabel = formatDistance(business.distanceKm);
-              const locationLabel = resolveListingLocationLabel(business.location, "brief") || "Nearby";
-              const imageSrc = toSafeImageSrc(business.coverImage || business.images?.[0], DEFAULT_IMAGE_PLACEHOLDER);
-              const businessIdentifier = (business.slug || business.id || "").toString().trim();
-              const businessHref = businessIdentifier ? `/business/${encodeURIComponent(businessIdentifier)}` : "/account/business";
-
-              return (
-                <Link
-                  key={business.id}
-                  href={businessHref}
-                  className="flex items-center gap-2.5 p-2 rounded-xl border border-slate-200/80 bg-white hover:border-blue-300 hover:bg-slate-50/60 transition-all group"
-                >
-                  <div className="relative size-10 shrink-0 rounded-lg overflow-hidden bg-slate-100 border border-slate-200/60">
-                    <SafeImage
-                      src={imageSrc}
-                      alt={business.name}
-                      fill
-                      unoptimized
-                      className="object-cover group-hover:scale-105 transition-transform"
-                      sizes="40px"
-                    />
-                  </div>
-
-                  <div className="flex-1 min-w-0">
-                    <h4 className="text-xs font-bold text-slate-800 group-hover:text-blue-600 truncate transition-colors">
-                      {business.name}
-                    </h4>
-                    <div className="flex items-center gap-1.5 text-2xs text-slate-500 mt-0.5">
-                      <span className="text-amber-500 font-semibold flex items-center">★ 4.5</span>
-                      <span>·</span>
-                      <span className="truncate">{locationLabel}</span>
-                      {distanceLabel ? <span>· {distanceLabel}</span> : null}
-                    </div>
-                  </div>
-                </Link>
-              );
-            })}
-          </div>
-        ) : null}
-
-        <div className="text-center pt-1">
-          <Link
-            href="/services"
-            className="text-xs font-semibold text-blue-600 hover:text-blue-700 transition-colors inline-flex items-center gap-1"
-          >
-            <span>View more services</span>
-            <ChevronRight className="h-3 w-3" />
-          </Link>
-        </div>
-      </div>
+      <RelatedBusinessSidebar
+        businesses={businesses}
+        isLoading={isLoading}
+        isError={isError}
+        title={sectionCopy.title}
+        description={sectionCopy.description}
+        emptyCopy={sectionCopy.empty}
+        formatDistance={formatDistance}
+      />
     );
   }
-
-  const renderCard = (business: Business) => {
-    const distanceLabel = formatDistance(business.distanceKm);
-    const matchingServicesCount = business.matchingServicesCount || 0;
-    const activeServicesCount = business.activeServicesCount || 0;
-    const locationLabel = resolveListingLocationLabel(business.location, "full") || "Nearby";
-    const imageSrc = toSafeImageSrc(business.coverImage || business.images?.[0], DEFAULT_IMAGE_PLACEHOLDER);
-    const businessIdentifier = (business.slug || business.id || "").toString().trim();
-    const businessHref = businessIdentifier ? `/business/${encodeURIComponent(businessIdentifier)}` : "/account/business";
-
-    return (
-      <Link key={business.id} href={businessHref} className="block shrink-0 group">
-        <Card className="w-56 md:w-60 shrink-0 border border-border shadow-2xs rounded-xl bg-card p-2.5 md:p-3 space-y-2 group-hover:border-primary/40 transition-colors">
-          <div className="flex items-start gap-2.5">
-            <div className="relative size-11 md:size-12 shrink-0 rounded-lg overflow-hidden bg-muted/50 border border-border">
-              <SafeImage
-                src={imageSrc}
-                alt={business.name}
-                fill
-                unoptimized
-                className="object-cover group-hover:scale-105 transition-transform duration-300"
-                sizes="48px"
-              />
-            </div>
-
-            <div className="flex-1 min-w-0">
-              <div className="flex items-center gap-1">
-                <h3 className="line-clamp-1 text-xs font-bold text-foreground group-hover:text-primary transition-colors flex-1">
-                  {business.name}
-                </h3>
-                {business.status === "live" && (
-                  <Badge className="shrink-0 rounded-full bg-blue-50 text-blue-700 px-1.5 py-0.5 text-2xs font-semibold border-none">
-                    Verified
-                  </Badge>
-                )}
-              </div>
-              <div className="flex items-center gap-1 text-tiny text-muted-foreground mt-0.5">
-                <MapPin className="h-3 w-3 shrink-0 text-muted-foreground/70" />
-                <span className="truncate">{locationLabel}</span>
-                {distanceLabel ? <span className="shrink-0">· {distanceLabel}</span> : null}
-              </div>
-            </div>
-          </div>
-
-          <div className="flex items-center justify-between gap-2 pt-0.5">
-            <div className="flex flex-wrap gap-1">
-              {matchingServicesCount > 0 ? (
-                <Badge variant="secondary" className="rounded-md bg-blue-50 px-1.5 py-0.5 text-2xs font-medium text-blue-700 border-none">
-                  {matchingServicesCount} matching
-                </Badge>
-              ) : activeServicesCount > 0 ? (
-                <Badge variant="secondary" className="rounded-md bg-slate-100 px-1.5 py-0.5 text-2xs font-medium text-slate-600 border-none">
-                  {activeServicesCount} live
-                </Badge>
-              ) : null}
-            </div>
-
-            <span className="inline-flex items-center justify-center h-7 px-2.5 rounded-md bg-blue-600 group-hover:bg-blue-700 text-white font-semibold text-2xs shrink-0 transition-colors">
-              <Wrench className="mr-1 h-3 w-3" />
-              View
-            </span>
-          </div>
-        </Card>
-      </Link>
-    );
-  };
 
   return (
     <section id="nearby-repair-services" className="mt-4 md:mt-6 px-3.5 md:px-0">
@@ -370,7 +191,13 @@ export function RelatedBusinessesSection({
           className="flex gap-4 overflow-x-auto pb-4 scrollbar-hide scroll-smooth"
           style={{ scrollbarWidth: "none", msOverflowStyle: "none" }}
         >
-          {businesses.map(renderCard)}
+          {businesses.map((business) => (
+            <RelatedBusinessCard
+              key={business.id}
+              business={business}
+              distanceLabel={formatDistance(business.distanceKm)}
+            />
+          ))}
         </div>
       ) : null}
     </section>
